requirements.txt
python-telegram-bot==21.1.1
black
pylint
mypy
isort
